third_party/WebKit/LayoutTests/fast/dom/HTMLFontElement/size-attribute-expected.txt

HTMLFontElement size attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fontSizeAttributeEffect("") is null
PASS fontSizeAttributeEffect("1") is "10px"
PASS fontSizeAttributeEffect("2") is "13px"
PASS fontSizeAttributeEffect("3") is "16px"
PASS fontSizeAttributeEffect("4") is "18px"
PASS fontSizeAttributeEffect("5") is "24px"
PASS fontSizeAttributeEffect("6") is "32px"
PASS fontSizeAttributeEffect("7") is "48px"
PASS fontSizeAttributeEffect("0") is "10px"
PASS fontSizeAttributeEffect("-1") is "13px"
PASS fontSizeAttributeEffect("-2") is "10px"
PASS fontSizeAttributeEffect("-3") is "10px"
PASS fontSizeAttributeEffect("-4") is "10px"
PASS fontSizeAttributeEffect("-5") is "10px"
PASS fontSizeAttributeEffect("-6") is "10px"
PASS fontSizeAttributeEffect("-7") is "10px"
PASS fontSizeAttributeEffect("-8") is "10px"
PASS fontSizeAttributeEffect("-9") is "10px"
PASS fontSizeAttributeEffect("-10") is "10px"
PASS fontSizeAttributeEffect("x6") is null
PASS fontSizeAttributeEffect(" 6") is "32px"
PASS fontSizeAttributeEffect("\t6") is "32px"
PASS fontSizeAttributeEffect("\r6") is "32px"
PASS fontSizeAttributeEffect("\n6") is "32px"
PASS fontSizeAttributeEffect("\u20086") is null
PASS fontSizeAttributeEffect("x-6") is null
PASS fontSizeAttributeEffect(" -6") is "10px"
PASS fontSizeAttributeEffect("\t-6") is "10px"
PASS fontSizeAttributeEffect("\r-6") is "10px"
PASS fontSizeAttributeEffect("\n-6") is "10px"
PASS fontSizeAttributeEffect("\u2008-6") is null
PASS fontSizeAttributeEffect("x+6") is null
PASS fontSizeAttributeEffect(" +6") is "48px"
PASS fontSizeAttributeEffect("\t+6") is "48px"
PASS fontSizeAttributeEffect("\r+6") is "48px"
PASS fontSizeAttributeEffect("\n+6") is "48px"
PASS fontSizeAttributeEffect("\u2008+6") is null
PASS fontSizeAttributeEffect("x+x6") is null
PASS fontSizeAttributeEffect(" + 6") is null
PASS fontSizeAttributeEffect("\t+\t6") is null
PASS fontSizeAttributeEffect("\r+\r6") is null
PASS fontSizeAttributeEffect("\n+\n6") is null
PASS fontSizeAttributeEffect("\u2008+\u20086") is null
PASS fontSizeAttributeEffect("x-x6") is null
PASS fontSizeAttributeEffect(" - 6") is null
PASS fontSizeAttributeEffect("\t-\t6") is null
PASS fontSizeAttributeEffect("\r-\r6") is null
PASS fontSizeAttributeEffect("\n-\n6") is null
PASS fontSizeAttributeEffect("\u2008-\u20086") is null
PASS fontSizeAttributeEffect("8") is "48px"
PASS fontSizeAttributeEffect("9") is "48px"
PASS fontSizeAttributeEffect("10") is "48px"
PASS fontSizeAttributeEffect("100") is "48px"
PASS fontSizeAttributeEffect("1000") is "48px"
PASS fontSizeAttributeEffect("1x") is "10px"
PASS fontSizeAttributeEffect("1.") is "10px"
PASS fontSizeAttributeEffect("1.9") is "10px"
PASS fontSizeAttributeEffect("2x") is "13px"
PASS fontSizeAttributeEffect("2.") is "13px"
PASS fontSizeAttributeEffect("2.9") is "13px"
PASS fontSizeAttributeEffect("a") is null
PASS fontSizeAttributeEffect(arabicIndicDigitOne) is null
PASS successfullyParsed is true

TEST COMPLETE
